modbus_bridge/requirements.txt
kafka-python==2.2.12
keyboard==0.13.5
pymodbus==3.9.2
